+++ drawFunc.hpp
@@ -11,7 +11,7 @@ void drawChristmasTreeObstacle(int position);
 void drawSteel(void);
 void drawSteelObstacle(int position);
 void drawObstacles(void);
-void drawObstacle(int x,int type,int position);
+void drawObstacle(float x,int type,int position);
 void BaltoCrashed(int firstBaltoObstacle);
 void randInitialization(void);
 void countdown(void);
